fix(당첨 상금): 당첨 상금이 int의 범위를 벗어나는 경우를 고려해 long 타입으로 변경

+++ src/main/java/lotto/controller/LottoController.java
@@ -52,7 +52,7 @@ public class LottoController {
     }
 
     private void printLottosProfitRate(PurchaseAmount purchaseAmount, Lottos lottos, WinningNumbers winningNumbers) {
-        int winningAmount = lottos.calculateWinningAmount(winningNumbers);
+        long winningAmount = lottos.calculateWinningAmount(winningNumbers);
         double profitRate = purchaseAmount.calculateProfitRate(winningAmount);
 
         outputView.printProfitRate(profitRate);
